Move get_app_context and is_rerunnable to app_context.py
